torchserve/tests/requirements.txt
pytest-docker==1.0.1
requests